Layers reorganized: two models for UI layer, mapping functions moved to extension functions in UI or Data layer | Tests updated with modifications | DetailViewModel created

+++ app/src/main/java/com/openclassrooms/vitesseapp/data/entity/CandidateDto.kt
@@ -30,7 +30,7 @@ data class CandidateDto(
     var birthdate: Long,
 
     @ColumnInfo(name = "salary_in_euro")
-    var salaryInEur: Int?,
+    var salaryInEur: Long?,
 
     @ColumnInfo(name = "notes")
     var notes: String?,
